- Finally some refactoring changes - My OCD can breathe a little easier for now, haha :P
