--- properties.py
@@ -1,15 +1,0 @@
-# Non-MacOS users can change it to Chrome/Firefox.
-BROWSER = "Safari" # can be Chrome/Safari/Firefox
-MATCH_URL = "http://www.espncricinfo.com/series/8039/commentary/1144506/afghanistan-vs-england-24th-match-icc-cricket-world-cup-2019"
-MESSAGE_BOX_CLASS_NAME = "_3u328"
-SEND_BUTTON_CLASS_NAME = "_3M-N-"
-
-# Match start timings according to where the script is being run.
-MATCH_START_HOURS = 13
-MATCH_START_MINUTES = 30
-MATCH_END_HOURS = 23
-MATCH_END_MINUTES = 0
-LOG_FILENAME = 'script_logs.log'
-
-# This tells whether the script is running in test mode or not. If yes, the actual data sent is less, for better debugging.
-IS_TEST_MODE = False
